Crate kube_runtime

Common components for building Kubernetes operators

This crate contains the core building blocks to allow users to build controllers/operators/watchers that need to synchronize/reconcile kubernetes state.

Newcomers are recommended to start with the Controller builder, which gives an opinionated starting point that should be appropriate for simple operators, but all components are designed to be usable á la carte if your operator doesn’t quite fit that mold.

Re-exports

pub use controller::applier;

pub use controller::Config;

pub use controller::Controller;

pub use finalizer::finalizer;

pub use reflector::reflector;

pub use scheduler::scheduler;

pub use utils::WatchStreamExt;

pub use watcher::metadata_watcher;

pub use watcher::watcher;

pub use utils::predicates;

pub use utils::Predicate;

pub use wait::conditions;

Modules

controller

Runs a user-supplied reconciler function on objects when they (or related objects) are updated

events

Publishes events for objects for kubernetes >= 1.19

finalizer

Finalizer helper for Controller reconcilers

reflector

Caches objects in memory

scheduler

Delays and deduplicates Stream items

utils

Helpers for manipulating built-in streams

wait

Waits for objects to reach desired states

watcher

Watches a Kubernetes Resource for changes, with error recovery