kube_runtime::wait

Module conditions

Common conditions to wait for

Re-exports

pub use super::Condition;

Structs

And

See Condition::and

Not

See Condition::not

Or

See Condition::or

Functions

is_crd_established

An await condition for CustomResourceDefinition that returns true once it has been accepted and established

is_deleted

An await condition that returns true once the object has been deleted.

is_deployment_completed

An await condition for Deployment that returns true once the latest deployment has completed

is_ingress_provisioned

An await condition for Ingress that returns true once the backing load balancer has an external IP or hostname

is_job_completed

An await condition for Job that returns true once it is completed

is_pod_running

An await condition for Pod that returns true once it is running

is_service_loadbalancer_provisioned

An await condition for Services of type LoadBalancer that returns true once the backing load balancer has an external IP or hostname