kube_runtime::wait::conditions

Function is_deployment_completed

pub fn is_deployment_completed() -> impl Condition<Deployment>

An await condition for Deployment that returns true once the latest deployment has completed

This looks for the condition that Kubernetes sets for completed deployments: https://kubernetes.io/docs/concepts/workloads/controllers/deployment/#complete-deployment