kube_runtime::wait::delete

Function delete_and_finalize

pub async fn delete_and_finalize<K: Clone + Debug + Send + DeserializeOwned + Resource + 'static>(
    api: Api<K>,
    name: &str,
    delete_params: &DeleteParams,
) -> Result<(), Error>

Delete an object, and wait for it to be removed from the Kubernetes API (including waiting for all finalizers to unregister themselves).

Errors

Returns an Error if the object was unable to be deleted, or if the wait was interrupted.