Module kube_runtime::controller

Runs a user-supplied reconciler function on objects when they (or related objects) are updated

Structs

Context

A context data type that’s passed through to the controllers callbacks

Controller

Controller

ReconcileRequest

A request to reconcile an object, annotated with why that request was made.

ReconcilerAction

Results of the reconciliation attempt

Enums

Error

ReconcileReason

Functions

applier

Apply a reconciler to an input stream, with a given retry policy

trigger_owners

Enqueues any owners of type KOwner for reconciliation

trigger_self

Enqueues the object itself for reconciliation

trigger_with

Helper for building custom trigger filters, see the implementations of trigger_self and trigger_owners for some examples.