Function kube_runtime::reflector::reflector

pub fn reflector<K, W>(
    store: Writer<K>, 
    stream: W
) -> impl Stream<Item = W::Item> where
    K: Resource + Clone,
    K::DynamicType: Eq + Hash + Clone,
    W: Stream<Item = Result<Event<K>>>, 

Caches objects from watcher::Events to a local Store

Keep in mind that the Store is just a cache, and may be out of date.

Note: It is a bad idea to feed a single reflector from multiple watchers, since the whole Store will be cleared whenever any of them emits a Restarted event.